requirements/dev.txt
-r base.txt

# needed only for testing app
# unneeded for production (handle in apache/nginx/etc.)
django-sslserver==0.19
